enh: delete icon changed to checkbox, selected items logic added, select objective code cleanup
